backend/requirements.txt
# Face Recognition Backend

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.26.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
pillow==10.2.0
dlib==19.24.4
